delete Auth Test, complete update, delete user

+++ shop/app/Traits/HandleUpdateImageTrait.php
@@ -22,7 +22,7 @@ trait HandleUpdateImageTrait
         }
     }
     public function deleteImage( $imageName){
-        if (file_exists($this->path.$imageName)) {
+        if ($imageName && file_exists($this->path.$imageName)) {
             unlink($this->path.$imageName);
         }
     }
